Refactor; Moved Task creation into TaskFactory
Improves SLAP so now TaskList is no longer responsible
for creating Task, and we can move the deserealising
of Task objects from Storage into TaskFactory.
